+++ src/Products/PloneMeeting/tests/testMeeting.py
@@ -2270,7 +2270,8 @@ class testMeetingType(PloneMeetingTestCase):
         item.REQUEST['PUBLISHED'] = item
         # as no current meeting and no meeting in the future, the item
         # may not be presented
-        self.assertFalse(item.wfConditions().mayPresent())
+        self.assertEqual(item.wfConditions().mayPresent().msg,
+                         u'not_able_to_find_meeting_to_present_item_into')
         # MeetingItem.getMeetingToInsertIntoWhenNoCurrentMeetingObject returns nothing
         # as no meeting in the future
         self.assertIsNone(item.getMeetingToInsertIntoWhenNoCurrentMeetingObject())